# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "small chat server",
        "title": "chat server",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/chats/add": {
            "post": {
                "description": "creates new chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "new chat",
                "parameters": [
                    {
                        "description": "Add chat",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddChat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ChatId"
                        }
                    }
                }
            }
        },
        "/chats/get": {
            "post": {
                "description": "get chat by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get chat",
                "parameters": [
                    {
                        "description": "user",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Chat"
                        }
                    }
                }
            }
        },
        "/messages/add": {
            "post": {
                "description": "sends new message to chat from user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "new message",
                "parameters": [
                    {
                        "description": "Add message",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageId"
                        }
                    }
                }
            }
        },
        "/messages/get": {
            "post": {
                "description": "get all messages from chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get messages",
                "parameters": [
                    {
                        "description": "chat id",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChatIdS"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Message"
                            }
                        }
                    }
                }
            }
        },
        "/users/add": {
            "post": {
                "description": "creates new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "new user",
                "parameters": [
                    {
                        "description": "Add user",
                        "name": "chat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserId"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddChat": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.AddMessage": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "integer"
                },
                "chat": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.AddUser": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        },
        "models.Chat": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "users_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.ChatId": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "integer"
                }
            }
        },
        "models.ChatIdS": {
            "type": "object",
            "properties": {
                "chat": {
                    "type": "integer"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "integer"
                },
                "chat": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.MessageId": {
            "type": "object",
            "properties": {
                "message_id": {
                    "type": "integer"
                }
            }
        },
        "models.UserId": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "integer"
                }
            }
        }
    }
}